CB-4489 - fixing network creation and deletion in env service

+++ cloud-azure/src/main/java/com/sequenceiq/cloudbreak/cloud/azure/AzureNetworkConnector.java
@@ -86,18 +86,21 @@ public class AzureNetworkConnector implements NetworkConnector {
 
     @Override
     public void deleteNetworkWithSubnets(NetworkDeletionRequest networkDeletionRequest) {
-        AzureClient azureClient = azureClientService.getClient(networkDeletionRequest.getCloudCredential());
-        try {
-            azureClient.deleteTemplateDeployment(networkDeletionRequest.getResourceGroup(), networkDeletionRequest.getStackName());
-            azureClient.deleteResourceGroup(networkDeletionRequest.getResourceGroup());
-        } catch (CloudException e) {
-            LOGGER.warn("Deletion error, cloud exception happened: ", e);
-            if (e.body() != null && e.body().details() != null) {
-                String details = e.body().details().stream().map(CloudError::message).collect(Collectors.joining(", "));
-                throw new CloudConnectorException(String.format("Stack deletion failed, status code %s, error message: %s, details: %s",
-                        e.body().code(), e.body().message(), details));
-            } else {
-                throw new CloudConnectorException(String.format("Stack deletion failed: '%s', please go to Azure Portal for detailed message", e));
+        if (!networkDeletionRequest.isExisting()) {
+            try {
+                AzureClient azureClient = azureClientService.getClient(networkDeletionRequest.getCloudCredential());
+                azureClient.deleteTemplateDeployment(networkDeletionRequest.getResourceGroup(), networkDeletionRequest.getStackName());
+                azureClient.deleteResourceGroup(networkDeletionRequest.getResourceGroup());
+
+            } catch (CloudException e) {
+                LOGGER.warn("Deletion error, cloud exception happened: ", e);
+                if (e.body() != null && e.body().details() != null) {
+                    String details = e.body().details().stream().map(CloudError::message).collect(Collectors.joining(", "));
+                    throw new CloudConnectorException(String.format("Stack deletion failed, status code %s, error message: %s, details: %s",
+                            e.body().code(), e.body().message(), details));
+                } else {
+                    throw new CloudConnectorException(String.format("Stack deletion failed: '%s', please go to Azure Portal for detailed message", e));
+                }
             }
         }
     }
